rename : user store 파일 이동 #26
