+++ src/app/calendar/page.tsx
@@ -1,23 +1,10 @@
-// CounterStore.tsx
-import {create} from "zustand";
 
-type State = {
-    isLogin: boolean;
+const page = () => {
+    return (
+        <div>
+            calendar
+        </div>
+    );
 };
 
-type Actions = {
-    userLogin: () => void;
-    userLogout: () => void;
-};
-
-const initialState: State = {
-    isLogin: false,
-};
-
-const useUserStore = create<State & Actions>((set) => ({
-    ...initialState,
-    userLogin: () => set((state) => ({ isLogin: true })),
-    userLogout: () => set((state) => ({ isLogin: false })),
-}));
-
-export default useUserStore;
+export default page;
